Version 8.1
New Inventory object/ Utilize a hash, and is a list of a pair and
anything, including arrays, objects.
Inventory is the only object in M2000  without copy method, and with
multiple reference to it. We can create global, local, or in a group.
Inventory deleted if not reference exist.
New is that we can use arrays as elements for arrays. In fact we can
bind arrays in groups and inventories and arrays, and we can bind groups
in arrays and inventories, and we can bind inventories to groups and
inventories.  Groups may have Groups inside. Arrays, Inventories and
Groups and variables may have lambda functions also.