Version 8.1 Revision 11
Addition for Functionality: Local command in a thread (in sequential
mode only) make not only a new variable but also mark all new variables
in the thread to deleted after a time interval execution. A thread can
be use Static Variables for private use or Module's variables for
exchange data. For inter thread communication, we can execute commands
in thread from other thread. So in that situation is better to send data
to thread stack or static variables.